+++ src/compress.ts
@@ -1,5 +1,6 @@
+import { createReadStream, createWriteStream } from 'fs'
 import type { Gzip } from 'zlib'
-import type { CompressionOptions, Algorithm } from './interface'
+import type { CompressionOptions, Algorithm, Compress } from './interface'
 
 export const ensureAlgorithmAndFormat = async (
   algorithm: Algorithm,
@@ -18,4 +19,16 @@ export const ensureAlgorithmAndFormat = async (
     default:
       throw new Error('Invalid algorithm')
   }
+}
+
+export const transfer = (entry: string, to: string, compress: Compress): Promise<number> => {
+  const len = []
+  return new Promise((resolve, reject) => {
+    createReadStream(entry)
+      .pipe(compress)
+      .on('data', (chunk) => len.push(chunk))
+      .pipe(createWriteStream(to))
+      .on('close', () => resolve(Buffer.concat(len).byteLength))
+      .on('error', reject)
+  })
 }
