+++ src/utils.ts
@@ -16,15 +16,6 @@ export const readGlobalFiles = async (entry: string, regular: Regular) => {
   return files
 }
 
-export const fromatBytes = (bytes: number, decimals = 2) => {
-  if (bytes === 0) return '0 Bytes'
-  const k = 1024
-  const dm = decimals < 0 ? 0 : decimals
-  const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB']
-  const i = Math.floor(Math.log(bytes) / Math.log(k))
-  return parseFloat((bytes / Math.pow(k, i)).toFixed(dm)) + ' ' + sizes[i]
-}
-
 export const len = <T>(source?: T[] | string) => source.length
 
 export const printf = {
